requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
numpy==1.25.1
pandas==2.0.3
python-dateutil==2.8.2   
python-dotenv==1.0.0     
pytz==2023.3
requests==2.31.0
six==1.16.0
tzdata==2023.3
urllib3==2.0.4